refactor: Identifies currency type
Identifies the type of currency that the user wants to convert to. The currency type is identified by the user's input. The user can input the currency type in the form of a string. The currency type is then stored in a variable. The variable is then used to identify the currency type in the conversion process.